animation updates
changed quat interpolate to slerp
split code into new files
serialize animation tree done
