+++ Editor/src/Ouroboros/Animation/Anim.h
@@ -14,6 +14,11 @@ Technology is prohibited.
 
 #pragma once
 #include "Anim_Utils.h"
+#include "AnimationGroup.h"
+#include "AnimationNode.h"
+#include "AnimationLink.h"
+#include "AnimationCondition.h"
+#include "AnimationParameter.h"
 #include "AnimationTree.h"
 #include "AnimationTracker.h"
 namespace oo::Anim::internal
