+++ Editor/src/Ouroboros/Animation/Anim_Utils.cpp
@@ -14,8 +14,24 @@ Technology is prohibited.
 #include "pch.h"
 #include "Anim_Utils.h"
 #include "Archetypes_Ecs/src/A_Ecs.h"
+#include "AnimationNode.h"
+#include "AnimationCondition.h"
 
 #include <rttr/instance.h>
+#include <rttr/registration>
+
+RTTR_REGISTRATION
+{
+	using namespace rttr;
+	registration::enumeration<oo::Anim::P_TYPE>("Parameter Type")
+	(
+		value("BOOL", oo::Anim::P_TYPE::BOOL),
+		value("TRIGGER", oo::Anim::P_TYPE::TRIGGER),
+		value("INT", oo::Anim::P_TYPE::INT),
+		value("FLOAT", oo::Anim::P_TYPE::FLOAT)
+	)
+	;
+}
 namespace oo::Anim::internal
 {
 	template<typename T>
@@ -71,69 +87,10 @@ namespace oo::Anim
 		return container;
 	}();
 
-	Condition::CompareFnMap Condition::comparisonFn_map = []()
-	{
-		rttr::type::register_comparators<bool>();
-		rttr::type::register_comparators<float>();
-		rttr::type::register_comparators<int>();
-
-		Condition::CompareFnMap map;
-		/*-------------------
-		BOOL / TRIGGER
-		-------------------*/
-		map[P_TYPE::BOOL][Condition::CompareType::EQUAL] =
-			[](DataType const& lhs, DataType const& rhs)
-		{
-			return lhs == rhs;
-		};
-		map[P_TYPE::TRIGGER][Condition::CompareType::EQUAL] =
-			[](DataType const& lhs, DataType const& rhs)
-		{
-			return lhs == rhs;
-		};
-
-		/*-------------------
-		FLOAT
-		-------------------*/
-		map[P_TYPE::FLOAT][Condition::CompareType::GREATER] =
-			[](DataType const& lhs, DataType const& rhs)
-		{
-			return lhs > rhs;
-		};
-		map[P_TYPE::FLOAT][Condition::CompareType::LESS] =
-			[](DataType const& lhs, DataType const& rhs)
-		{
-			return lhs < rhs;
-		};
-		/*-------------------
-		INT
-		-------------------*/
-		map[P_TYPE::INT][Condition::CompareType::EQUAL] =
-			[](DataType const& lhs, DataType const& rhs)
-		{
-			return lhs == rhs;
-		};
-		map[P_TYPE::INT][Condition::CompareType::NOT_EQUAL] =
-			[](DataType const& lhs, DataType const& rhs)
-		{
-			return lhs != rhs;
-		};
-		map[P_TYPE::INT][Condition::CompareType::GREATER] =
-			[](DataType const& lhs, DataType const& rhs)
-		{
-			return lhs > rhs;
-		};
-		map[P_TYPE::INT][Condition::CompareType::LESS] =
-			[](DataType const& lhs, DataType const& rhs)
-		{
-			return lhs < rhs;
-		};
-
-		return map;
-	}();
+	
 
 
-	Parameter::SerializeFnMap Parameter::serializeFn_map = 
+	/*Parameter::SerializeFnMap const Parameter::serializeFn_map = 
 		[]() {
 		Parameter::SerializeFnMap map;
 		
@@ -144,7 +101,247 @@ namespace oo::Anim
 			writer.Bool(param.value.get_value<bool>());
 		};
 
-		return map;
-		}();
+		map[P_TYPE::TRIGGER] =
+			[](rapidjson::PrettyWriter<rapidjson::OStreamWrapper>& writer, Parameter& param)
+		{
+			writer.String("TRIGGER", static_cast<rapidjson::SizeType>(std::string("TRIGGER").size()));
+			writer.Bool(param.value.get_value<bool>());
+		};
 
+		map[P_TYPE::FLOAT] =
+			[](rapidjson::PrettyWriter<rapidjson::OStreamWrapper>& writer, Parameter& param)
+		{
+			writer.String("FLOAT", static_cast<rapidjson::SizeType>(std::string("FLOAT").size()));
+			writer.Double(static_cast<double>(param.value.get_value<float>()));
+		};
+
+		map[P_TYPE::INT] =
+			[](rapidjson::PrettyWriter<rapidjson::OStreamWrapper>& writer, Parameter& param)
+		{
+			writer.String("INT", static_cast<rapidjson::SizeType>(std::string("INT").size()));
+			writer.Int(param.value.get_value<int>());
+		};
+
+		return map;
+		}();*/
+
+}
+
+namespace oo::Anim
+{
+	void SerializeGroupRef(rapidjson::PrettyWriter<rapidjson::OStreamWrapper>& writer, GroupRef& ref)
+	{
+		writer.StartObject();
+		if (ref)
+			writer.Key(ref->name.c_str(), static_cast<rapidjson::SizeType>(ref->name.size()));
+		else
+		{
+			LOG_CORE_DEBUG_CRITICAL("Group reference {0} invalid when serializing!!", ref.id);
+			assert(false);
+		}
+		writer.Uint64(ref.id);
+		writer.EndObject();
+	}
+	void SerializeNodeRef(rapidjson::PrettyWriter<rapidjson::OStreamWrapper>& writer, NodeRef& ref)
+	{
+		writer.StartObject();
+		if (ref)
+			writer.Key(ref->name.c_str(), static_cast<rapidjson::SizeType>(ref->name.size()));
+		else
+		{
+			LOG_CORE_DEBUG_CRITICAL("Node reference {0} invalid when serializing!!", ref.id);
+			assert(false);
+		}
+		writer.Uint64(ref.id);
+		writer.EndObject();
+	}
+	void SerializeLinkRef(rapidjson::PrettyWriter<rapidjson::OStreamWrapper>& writer, LinkRef& ref)
+	{
+		writer.StartObject();
+		if (ref)
+			writer.Key(ref->name.c_str(), static_cast<rapidjson::SizeType>(ref->name.size()));
+		else
+		{
+			LOG_CORE_DEBUG_CRITICAL("Node reference {0} invalid when serializing!!", ref.id);
+			assert(false);
+		}
+		writer.Uint64(ref.id);
+		writer.EndObject();
+	}
+	void SerializeAnimRef(rapidjson::PrettyWriter<rapidjson::OStreamWrapper>& writer, AnimRef& ref)
+	{
+		writer.StartObject();
+		if (ref)
+			writer.Key(ref->name.c_str(), static_cast<rapidjson::SizeType>(ref->name.size()));
+		else
+		{
+			LOG_CORE_DEBUG_CRITICAL("Node reference {0} invalid when serializing!!", ref.id);
+			assert(false);
+		}
+		writer.Uint64(ref.id);
+		writer.EndObject();
+	}
+	RTTR_REGISTRATION
+	{
+		using namespace rttr;
+		registration::class_<GroupRef>("Animation Group Reference")
+			.property("ID", &GroupRef::id)
+			.method(internal::serialize_method_name,&SerializeGroupRef)
+			;
+		registration::class_<NodeRef>("Animation Node Reference")
+			.property("ID", &NodeRef::id)
+			.method(internal::serialize_method_name, &SerializeNodeRef)
+			;
+		registration::class_<LinkRef>("Animation Link Reference")
+			.property("ID", &LinkRef::id)
+			.method(internal::serialize_method_name, &SerializeLinkRef)
+			;
+		registration::class_<AnimRef>("Animation Anim Reference")
+			.property("ID", &AnimRef::id)
+			.method(internal::serialize_method_name, &SerializeAnimRef)
+			;
+	};
+	GroupRef::GroupRef(std::map<size_t, Group>* _groups, size_t _id) :
+		groups{ _groups },
+		id{ _id }
+	{
+
+	}
+	void GroupRef::Reload()
+	{
+		assert(groups);
+		assert((*groups).contains(id));
+		/*int currindex{ 0 };
+		for (auto& group : *groups)
+		{
+			if (group.groupID == id)
+			{
+				index = currindex;
+				return;
+			}
+			++currindex;
+		}
+
+		LOG_CORE_ERROR("Animation Group Reference Reload failed!!");
+		assert(false);
+		index = -1;*/
+	}
+	bool GroupRef::valid() const
+	{
+		/*return groups && index >= 0 && index < groups->size() &&
+			this->operator->()->groupID == id;*/
+		assert(groups);
+		return (*groups).contains(id);
+	}
+	Group& GroupRef::Retrieve(size_t id) const
+	{
+		return (*groups).at(id);
+	}
+	/*-------------------------------
+	NodeRef
+	-------------------------------*/
+	NodeRef::NodeRef(std::map<size_t, Node>* _nodes, size_t _id) :
+		nodes{ _nodes },
+		id{ _id }
+	{
+
+	}
+
+	void NodeRef::Reload()
+	{
+		assert(nodes);
+		assert((*nodes).contains(id));
+		/*int currindex{0};
+		for (auto& [node_id, node] : *nodes)
+		{
+			if (node.node_ID == id)
+			{
+				index = currindex;
+				return;
+			}
+			++currindex;
+		}
+
+		LOG_CORE_ERROR("Animation Node Reference Reload failed!!");
+		assert(false);
+		index = -1;*/
+	}
+	bool NodeRef::valid() const
+	{
+		/*return nodes && index >= 0 && index < nodes->size() &&
+			this->operator->()->node_ID == id;*/
+		assert(nodes);
+		return (*nodes).contains(id);
+	}
+	Node& NodeRef::Retrieve(size_t id) const
+	{
+		return (*nodes).at(id);
+	}
+	/*-------------------------------
+	LinkRef
+	-------------------------------*/
+	LinkRef::LinkRef(std::map<size_t, Link>* _links, size_t _id) :
+		links{ _links },
+		id{ _id }
+	{
+
+	}
+	void LinkRef::Reload()
+	{
+		assert(links);
+		assert((*links).contains(id));
+		/*int currindex{ 0 };
+		for (auto& link : *links)
+		{
+			if (link.linkID == id)
+			{
+				index = currindex;
+				return;
+			}
+			++currindex;
+		}
+
+		LOG_CORE_ERROR("Animation Link Reference Reload failed!!");
+		assert(false);
+		index = -1;*/
+	}
+	bool LinkRef::valid() const
+	{
+		/*return links && index >= 0 && index < links->size() &&
+			this->operator->()->linkID == id;*/
+		assert(links);
+		return (*links).contains(id);
+	}
+	Link& LinkRef::Retrieve(size_t id) const
+	{
+		return (*links).at(id);
+	}
+	/*-------------------------------
+	AnimRef
+	-------------------------------*/
+	void AnimRef::Reload()
+	{
+		assert(anims);
+		assert((*anims).contains(id));
+		/*int currindex{ 0 };
+		for (auto& anim : *anims)
+		{
+			if (anim.animation_ID == id)
+			{
+				index = currindex;
+				return;
+			}
+			++currindex;
+		}
+
+		LOG_CORE_ERROR("Animation Reference Reload failed!!");
+		assert(false);
+		index = -1;*/
+	}
+	bool AnimRef::valid() const
+	{
+		/*return anims && index >= 0 && index < anims->size() &&
+			this->operator->()->animation_ID == id;*/
+		return (*anims).contains(id);
+	}
 }
