+++ Editor/src/Ouroboros/Scene/RuntimeScene.cpp
@@ -51,7 +51,7 @@ namespace oo
 
             GetWorld().Add_System<PhysicsSystem>()->Init(this);
 
-            //GetWorld().Get_System<Anim::AnimationSystem>()->CreateAnimationTestObject();
+            GetWorld().Get_System<Anim::AnimationSystem>()->CreateAnimationTestObject();
 
             GetWorld().Get_System<Anim::AnimationSystem>()->BindPhase();
             //Register All Systems
